requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
azure-ai-documentintelligence==1.0.0b3
openai>=1.3.0
azure-search-documents==11.4.0
tiktoken
reportlab
